requirements.txt
fastapi
uvicorn
streamlit
pandas
openpyxl
requests
openai
python-multipart
groq
load_dotenv